requirement.txt
asgiref==3.5.0
astroid==2.9.3
atomicwrites==1.4.0
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
Django==4.0.3
dominate==2.6.0
Flask==2.0.2
Flask-Bootstrap4==4.0.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.10.1
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
mysql-connector-python==8.0.28
packaging==21.3
platformdirs==2.4.1
pluggy==1.0.0
protobuf==3.19.4
py==1.11.0
PyLin==0.2.7
pylint==2.12.2
pyparsing==3.0.7
pyserial==3.5
requests==2.27.1
SQLAlchemy==1.4.32
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.0
typing-extensions==4.0.1
urllib3==1.26.8
visitor==0.1.3
Werkzeug==2.0.2
wrapt==1.13.3
WTForms==3.0.1
zipp==3.7.0
